The tree builds off Windows, and that is how R11d's step was measured and refused
Three GCC diagnostics were all that stood between this repository and a build on a
machine with no card, and the handover has been telling sessions the opposite since
D643. Two report helpers in crash.cpp have no caller off Windows; field.cpp's
distance_to has been dead since D638 and MSVC cannot see an unused function in an
anonymous namespace; node_pool.cpp mixes an enum and a u8 in one conditional. Plus
-Wno-missing-field-initializers for GCC only, which is 316 hits on Vulkan structs
written as designated initialisers and not one of them a fault.

So the whole headless clip tool now runs on Linux, and D647's next step could be
measured rather than argued about. --stipple-tiled tiles a clip into the node boxes
the ladder samples and counts them three ways against one whole-clip reference:

  bare, what ships today          50 materials DIFFER, 35 never seen   3.31 ms a node
  inner, free                      2 DIFFER, and both are dithers      3.31 ms a node
  skirted, a box one voxel larger  0 DIFFER, counts EXACT              6.69 ms a node

The skirted sum is not close, it is identical -- 81 of 81 materials, surface and
specks off by zero -- and its verdict is the shipped 27 358 392 455 509 554 material
for material. So D629's method holds at the ladder's own geometry, which is what R11d
needed. The price is what refutes the step: 2.02x at metre 8 and 2.13x at the authored
metre 32, so doubling the ladder's 6,322 ms to remove a 2,754 ms bar is three and a
half seconds the wrong way, paid on the nodes the player is waiting for. What survives
is the route D628 named and nobody has built -- read the skirt out of the WORLD after
the paste, where a neighbour costs a memory read and not a field evaluation.

A unit test pins the arithmetic with no sampler in the way, and keeps the bare arm
broken on purpose: a control that stops being wrong has stopped controlling anything.

D648, D649. 543 of 545 tests pass; the two failures are the recycle bin, which is a
Windows shell call.

+++ src/world/node_pool.cpp
@@ -1259,8 +1259,8 @@ const NodeUploadBatch& NodePool::update(const World& world, const f64 camera_vox
     for (usize asked = 0; asked < requested_.size(); ++asked) {
         const NodeKey& key = requested_[asked];
         if (!seen_requests_.insert(key).second) continue;
-        const u8 source =
-            (asked < requested_source_.size()) ? requested_source_[asked] : kRequestRay;
+        const u8 source = (asked < requested_source_.size()) ? requested_source_[asked]
+                                                             : static_cast<u8>(kRequestRay);
 
         // Instrument: is this a node this pool threw away a moment ago?
         //
